Add RestService for querying statestore

+++ src/main/java/annoyED/store/Datapoint.java
@@ -6,12 +6,12 @@ public class Datapoint {
 
 
   public String datapointID;
-  public Vector<Float> vector;
+  public Vector<Double> vector;
 
   public Datapoint() {
-    this("", new Vector<Float>());
+    this("", new Vector<Double>());
   }
-  public Datapoint(String datapointId, Vector<Float> vector) {
+  public Datapoint(String datapointId, Vector<Double> vector) {
     this.datapointID = datapointId;
     this.vector = vector;
   }
